+++ cnn.py
@@ -5,27 +5,41 @@ from pandas_datareader import data
 import tensorflow as tf
 import numpy as np
 
+# Define some macros
+TRAINING_SET_RATIO = 0.8
+
 #apple = Share('AAPL')
 #past_price = apple.get_historical('2016-05-01', '2016-05-31')
-training_test_data = data.DataReader('AAPL', 'yahoo', '2000-01-01')
+training_test_data = data.DataReader('AAPL', 'yahoo', '2016-07-01')
 
 training_test_data['Class'] = 0
+
+# reset the original indexing (by date) to 0..n
 training_test_data=training_test_data.reset_index(drop=True)
+
+# setting up some useful variables
+data_length = len(training_test_data['Open'])
+
+# set the classification data
 for i in range(0, len(training_test_data['Open'])):
     if training_test_data['Close'][i] > training_test_data['Open'][i]:
         training_test_data.set_value(i, 'Class', 1)
     else:
         training_test_data.set_value(i, 'Class', 0)
+
 for i in range(0, len(training_test_data['Open'])):
     training_test_data.set_value(i, 'Volume', training_test_data['Volume'][i]/1000000.0)
 
+# Remove the close values
 training_test_data = training_test_data.drop('Close', 1)
 training_test_data = training_test_data.drop('Adj Close', 1)
 
+# Split the data into x_i and y, in other words, features and results
 predictors_tf = training_test_data[training_test_data.columns[:-1]]
 classes_tf = training_test_data[training_test_data.columns[-1]]
 
-training_set_size = int(len(training_test_data)*0.8)
+# divide the training set into training and testing
+training_set_size = int(len(training_test_data)*TRAINING_SET_RATIO)
 test_set_size = len(training_test_data) - training_set_size
 
 training_predictors_tf = predictors_tf[:training_set_size]
@@ -33,6 +47,7 @@ training_classes_tf = classes_tf[:training_set_size]
 test_predictors_tf = predictors_tf[training_set_size:]
 test_classes_tf = classes_tf[training_set_size:]
 
+# Make tensorFlow session
 sess = tf.Session()
 
 num_predictors = len(training_predictors_tf.columns)
